+++ src/com/amazon/corretto/crypto/provider/AmazonCorrettoCryptoProvider.java
@@ -47,7 +47,7 @@ public final class AmazonCorrettoCryptoProvider extends java.security.Provider {
     private transient SelfTestSuite selfTestSuite = new SelfTestSuite();
 
     static {
-        if (!Loader.IS_AVAILABLE) {
+        if (!Loader.IS_AVAILABLE && DebugFlag.VERBOSELOGS.isEnabled()) {
             getLogger("AmazonCorrettoCryptoProvider").fine("Native JCE libraries are unavailable - disabling");
             rdRandSupported_ = false;
         } else {
@@ -220,7 +220,7 @@ public final class AmazonCorrettoCryptoProvider extends java.security.Provider {
                  * end up falling back.
                  */
                 if (!getType().equals("SecureRandom")) {
-                    throw new NoSuchAlgorithmException("Can't use AACP before JAR validation completes");
+                    throw new NoSuchAlgorithmException("Can't use ACCP before JAR validation completes");
                 }
             }
 
@@ -276,23 +276,9 @@ public final class AmazonCorrettoCryptoProvider extends java.security.Provider {
     public AmazonCorrettoCryptoProvider() {
         super("AmazonCorrettoCryptoProvider", PROVIDER_VERSION, "");
 
-        final String[] extraCheckOptions = Loader.getProperty("extrachecks", "").split(",");
-        for (final String check : extraCheckOptions) {
-          if (check.equalsIgnoreCase("all")) {
-            extraChecks.addAll(EnumSet.allOf(ExtraCheck.class));
-            break;
-          }
-          try {
-            final ExtraCheck value = ExtraCheck.valueOf(check.toUpperCase());
-            if (value != null) {
-              extraChecks.add(value);
-            }
-          } catch (Exception ex) {
-            // Ignore
-          }
-        }
+        Utils.optionsFromProperty(ExtraCheck.class, extraChecks, "extrachecks");
 
-        if (!Loader.IS_AVAILABLE) {
+        if (!Loader.IS_AVAILABLE && DebugFlag.VERBOSELOGS.isEnabled()) {
             getLogger("AmazonCorrettoCryptoProvider").fine("Native JCE libraries are unavailable - disabling");
 
             // Don't implement anything
@@ -347,7 +333,7 @@ public final class AmazonCorrettoCryptoProvider extends java.security.Provider {
      * {@link SelfTestStatus#NOT_RUN} will be returned if any tests have not be run.
      * {@link SelfTestStatus#PASSED} will only be returned if all tests have been run and have
      * all passed.
-     * 
+     *
      * <p>Algorithms currently run by this method:
      * <ul>
      * <li>NIST800-90A/AES-CTR-256
@@ -357,7 +343,7 @@ public final class AmazonCorrettoCryptoProvider extends java.security.Provider {
      * <li>HMacSHA1
      * <li>HMacMD5
      * </ul>
-     * 
+     *
      * @see #runSelfTests()
      */
     public SelfTestStatus getSelfTestStatus() {
@@ -369,7 +355,7 @@ public final class AmazonCorrettoCryptoProvider extends java.security.Provider {
      * Please see {@link #getSelfTestStatus()} for the algorithms tested and
      * the possible return values. (though this method will never return
      * {@link SelfTestStatus#NOT_RUN}).
-     * 
+     *
      * @see #getSelfTestStatus()
      */
     public SelfTestStatus runSelfTests() {
@@ -387,7 +373,7 @@ public final class AmazonCorrettoCryptoProvider extends java.security.Provider {
     /**
      * <p>Throws an instance of {@link RuntimeCryptoException} if this library is not currently
      * functional. Otherwise does nothing.
-     * 
+     *
      * <p>This library is considered healthy if {@link #getLoadingError()} returns {@code null}
      * and {@link #runSelfTests()} returns {@link SelfTestStatus#PASSED}.
      */