+++ CHANGELOG.md
@@ -8,6 +8,23 @@
 
 ### Patches
 * Correctly reject non-empty `PSource.PSpecified` values for RSA-OAEP.
+## 1.7.0 (Unreleased)
+### Improvements
+* Now uses [OpenSSL 1.1.1m](https://www.openssl.org/source/openssl-1.1.1m.tar.gz). [PR #173](https://github.com/corretto/amazon-corretto-crypto-provider/pull/173)
+* Add "help" value to two of our properties which outputs (to STDERR) valid values.
+   * `com.amazon.corretto.crypto.provider.extrachecks`
+   * `com.amazon.corretto.crypto.provider.debug`
+* Add new `com.amazon.corretto.crypto.provider.debug` property to gate possibly expensive debug logic.
+Current values are:
+   * `FreeTrace` - Enables tracking of allocation and freeing of native objects from java for more detailed exceptions.
+   * `VerboseLogging` - Enables more detailed logging.
+   * `ALL` - Enables all of the above
+(May still require changes to your logging configuration to see the new logs.)
+* Enables skipping the bundled lib by setting the system property `com.amazon.corretto.crypto.provider.useExternalLib` [PR #168](https://github.com/corretto/amazon-corretto-crypto-provider/pull/168)
+
+### Patches
+* Improve zeroization of DRBG output. [PR #162](https://github.com/corretto/amazon-corretto-crypto-provider/pull/162)
+
 ## 1.6.1
 ### Patches
 * Fix an issue where a race condition can cause ACCP's MessageDigest hashing algorithms to return the same value for different inputs [PR #157](https://github.com/corretto/amazon-corretto-crypto-provider/pull/157)